Migrate logical plan tests
